Add Tests Cases for the permutation, combination and factorial functions

+++ CompStructs/StatsMath.h
@@ -5,9 +5,9 @@ namespace CompStructs
 	static class StatsMath
 	{
 	public:
-		static long Combination(long n, long r, bool repeat);
-		static long Permutation(long n, long r, bool repeat);
-		static long Factorial(long n);
+		static unsigned long long Combination(unsigned long long n, unsigned long long r, bool repeat);
+		static unsigned long long Permutation(unsigned long long n, unsigned long long r, bool repeat);
+		static unsigned long long Factorial(unsigned long long n);
 	};
 }
